Move `src/examples/` as top-level `examples/` folder
Slowly converting to `pages/`
