+++ pages/accordions/index.js
