+++ pages/example.js
