+++ pages/tabs/index.js
